Desktop/Zoning-Project(NEW)/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# AI/ML
openai==1.3.7
requests==2.31.0

# Security and Rate Limiting
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# Redis (for caching/rate limiting)
redis==5.0.1

# Logging and Monitoring
structlog==23.2.0
